Split shared workspace helpers from app entrypoint
