+++ tools/controlplane/src/controlplane_tool/infra/runtimes/mockk8s_runtime.py
@@ -9,7 +9,7 @@ import httpx
 from tenacity import RetryError, Retrying, stop_after_delay, wait_fixed
 
 from controlplane_tool.core.process_streaming import spawn_logged_process
-from controlplane_tool.app.settings import ToolSettings
+from controlplane_tool.workspace.settings import ToolSettings
 
 
 @dataclass
