+++ tools/controlplane/src/controlplane_tool/infra/runtimes/prometheus_runtime.py
@@ -7,7 +7,7 @@ import httpx
 from tenacity import RetryError, Retrying, stop_after_delay, wait_fixed
 
 from controlplane_tool.core.shell_backend import ShellExecutionResult, SubprocessShell
-from controlplane_tool.app.settings import ToolSettings
+from controlplane_tool.workspace.settings import ToolSettings
 
 
 @dataclass(frozen=True)
