+++ tools/controlplane/src/controlplane_tool/infra/runtimes/registry_runtime.py
@@ -13,7 +13,7 @@ import time
 
 from controlplane_tool.scenario.scenario_runtime import wait_for_http_ok
 from controlplane_tool.core.shell_backend import ShellBackend, ShellExecutionResult, SubprocessShell
-from controlplane_tool.app.settings import ToolSettings
+from controlplane_tool.workspace.settings import ToolSettings
 
 
 def default_registry_url() -> str:
