+++ tools/controlplane/src/controlplane_tool/infra/vm/ansible_adapter.py
@@ -5,7 +5,7 @@ from typing import Callable
 
 from multipass import MultipassClient
 
-from controlplane_tool.app.paths import ToolPaths
+from controlplane_tool.workspace.paths import ToolPaths
 from controlplane_tool.core.shell_backend import (
     ShellBackend,
     ShellExecutionResult,
