+++ tools/controlplane/src/controlplane_tool/loadtest/loadtest_flows.py
@@ -12,7 +12,7 @@ from controlplane_tool.loadtest.loadtest_tasks import (
     run_loadtest_step_task,
     write_loadtest_report_task,
 )
-from controlplane_tool.app.paths import default_tool_paths
+from controlplane_tool.workspace.paths import default_tool_paths
 from controlplane_tool.orchestation.prefect_models import LocalFlowDefinition
 from controlplane_tool.core.run_models import RunResult
